requirements.txt
google-genai
watchdog
pystray
pillow
pyautogui
packaging
customtkinter
darkdetect
pyinstaller
